LayoutTests/imported/w3c/web-platform-tests/css/css-inline/baseline-source/baseline-source-vertical-align-expected.txt


FAIL baseline-source-vertical-align assert_equals: expected (string) "auto" but got (undefined) undefined
